CLI improvements

+++ index.js
@@ -1,28 +1,37 @@
 #!/usr/bin/env node
 
+const fs = require('fs');
 const path = require('path');
 const chokidar = require('chokidar');
-const { fork } = require('child_process');
+const { fork, spawn } = require('child_process');
 
 module.exports = function(paths, opts) {
   if (!process.send) {
-    let child, mod;
-    if (require.main === module) {
-      paths = path.dirname(process.argv[2]);
-      mod = path.join(process.cwd(), process.argv[2]);
-    } else if (!paths) {
-      paths = path.dirname(process.argv[1]);
+    let child, start;
+    if (require.main === module) {  // CLI
+      let bin = process.argv[2];
+      if (fs.existsSync(bin))
+        bin = path.join(paths, bin);
+      start = function() {
+        if (child) child.kill();
+        child = spawn(bin, process.argv.slice(3), {
+          stdio: 'inherit',
+        });
+      }
+    } else {  // require()
+      if (!paths)
+        paths = path.dirname(process.argv[1]);
+      start = function() {
+        if (child) child.kill();
+        child = fork(process.argv[1], process.argv.slice(2));
+      }
     }
-    const start = function() {
-      if (child)
-        child.kill();
-      child = fork(mod || process.argv[1]);
-    };
     chokidar.watch(paths, opts).on('ready', start).on('change', start);
   }
   return !process.send;
 }
 
-if (require.main === module) {
-  module.exports();
+if (require.main === module) {  // CLI
+  if (!process.argv[2]) throw "must pass argument!";
+  module.exports(process.cwd());
 }
